requirements.txt
requests
telebot
google-generativeai
markdownify
bs4
smtplib
